Make sure blockUpdates doesn't skip blocks
Sometimes (especially with the Catalyst engine) NewHead gets called
without being called for all intermediate blocks since the last block
it was called on. Before supporting the merge, this wasn't an issue
as it only occurred because of Reorgs, which were handled separately
through Reorg() calls.

But in some cases with the beacon chain, we may receive several blocks
at once, and skip straight to the most recent at the end.

This change recursively backs up to a recently emitted block, then processes
all intermediate blocks.