Peer Management Plugin
Back in the EtherCattle days, I wrote this script:

https://github.com/openrelayxyz/ethercattle-deployment/blob/master/devops/peerManager.py

Which had two intended functions:

1. Make sure all of our servers for a given network were peered with eachother
2. Drop peers that are not making forward progress

It turns out it never did #2 correctly, but also it doesn't work on
current versions of Python. This plugin incorporates function #1 to
make sure that Rivet's Geth nodes for a given chain are peered with
one another.

At some point, we might add hooks that will make function #2 possible,
but as it stands you can't do that with publicly exposed RPC methods.